[SOIN] Déplace la ressource d'informations de création de compte dans un fichier dédié
